extract table row to component
